Parse battles in ed7 scenas
This parsing is largely based on heuristics. They hold for all vanilla
scripts, but are not tested on EDDecompiler. I might relax the
first==100 heuristic on AT rolls later.

There exists at least one scena (r2000) that has rolls and placements
but no battles. This means those have to be distinct entities for
roundtripping purposes, and it also implies that they are distinct in
the source files.