Find function end with flow analysis
In Sky, the code chunk is immediately followed by the function table,
which means there is a reliable way to tell the end of it. This is
unfortunately not the case in Zero; the code chunk is followed by some
data that I think has to do with sepith drops (though this is
unverified).

In this case I instead check all jumps to find the extent of the last
function. This is not entirely reliable — Sky has a number of
unreferenced code blocks after the last function. But I haven't found
any such in Zero yet, so until I do this is enough. No idea how it will
work with EDDecompiler (notably Geofront), though.