Support triple-return
I guess the code there is `if(...) { ...; return } return`, plus the
implicit return by the compiler. This implicit return is not found by
the flow checker, so I have to add a special case to handle it. Still
can't handle if there's actual code after an unconditional return in the
last function, but none of the scripts have that.