Decompile most EDDecompiled functions
EDDec messes up the control flow a lot, mainly by removing unreachable gotos, such as

```
if(...) {
	return
} else {
	(stuff)
}
```

This corresponds to the instruction sequence

```
Unless ... L1
Return
Goto L2
@L1
(stuff)
@L2
```

However, EDDecompiler notices that the `Goto L2` is unreachable and removes it, which makes the code indistinguishable from

```
if(...) {
	return
}
(stuff)
```

The way to work around this is to try to copy over the removed
instructions from the vanilla script. This is of course very unreliable
and distasteful, but it works on most of the scripts in SoraVoice.

There are however many cases where it does not work: I've seen a few
places where a basic block exists in EDDec, but at the wrong position.
And it will not work if the control flow is *actually* changed, such as
realphabetizing the Crossbell library or Sully's retry.