Move current stuff to themelios-scena
I plan to move out as much as possible into a separate crate from the
code parsing, because compiling that macro invocation and huge enum is
slow.

Next step is to move the testing to a separate crate, so that it can use
as many dependencies as it needs. After this more things will be moved
out into separate crates.

Probably the final result will be to have a `common` crate at the bottom
defining various types, traits, and utils, which is used by `table` and
`scena` crates, which in turn are used by the top-level `themelios`
module, which also handles IO, and is the one intended for public use.